[DAD-776] refactor: ScrapController 폴더 구조 를 각각의 카테고리 Controller로 변경
