+++ src/main/java/com/forever/dadamda/dto/scrap/SearchScrapRequest.java
@@ -1,0 +1,8 @@
+package com.forever.dadamda.dto.scrap;
+
+import lombok.Getter;
+
+@Getter
+public class SearchScrapRequest {
+
+}
